Split various common functionality into semantic interfaces
