+++ src/Contracts/Authenticatable.php
@@ -1,0 +1,60 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Saloon\Contracts;
+
+interface Authenticatable
+{
+    /**
+     * Retrieve the authenticator.
+     *
+     * @return \Saloon\Contracts\Authenticator|null
+     */
+    public function getAuthenticator(): ?Authenticator;
+
+    /**
+     * Authenticate the request with an authenticator.
+     *
+     * @param \Saloon\Contracts\Authenticator $authenticator
+     * @return $this
+     */
+    public function authenticate(Authenticator $authenticator): static;
+
+    /**
+     * Authenticate the request with an Authorization header.
+     *
+     * @param string $token
+     * @param string $prefix
+     * @return $this
+     */
+    public function withTokenAuth(string $token, string $prefix = 'Bearer'): static;
+
+    /**
+     * Authenticate the request with "basic" authentication.
+     *
+     * @param string $username
+     * @param string $password
+     * @return $this
+     */
+    public function withBasicAuth(string $username, string $password): static;
+
+    /**
+     * Authenticate the request with "digest" authentication.
+     *
+     * @param string $username
+     * @param string $password
+     * @param string $digest
+     * @return $this
+     */
+    public function withDigestAuth(string $username, string $password, string $digest): static;
+
+    /**
+     * Authenticate the request with a query parameter token.
+     *
+     * @param string $parameter
+     * @param string $value
+     * @return $this
+     */
+    public function withQueryAuth(string $parameter, string $value): static;
+}
