+++ src/Contracts/Makeable.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Saloon\Contracts;
+
+interface Makeable
+{
+    /**
+     * Instantiate a new class with the arguments.
+     *
+     * @param mixed ...$arguments
+     * @return static
+     */
+    public static function make(mixed ...$arguments): static;
+}
