+++ src/Contracts/PendingRequest.php
@@ -8,130 +8,8 @@ use Saloon\Enums\Method;
 use GuzzleHttp\Promise\PromiseInterface;
 use Saloon\Contracts\Body\BodyRepository;
 
-interface PendingRequest
+interface PendingRequest extends Authenticatable, Conditionable, HasConfig, HasHeaders, HasMiddlewarePipeline, HasMockClient, HasQueryParams
 {
-    /**
-     * Retrieve the authenticator.
-     *
-     * @return Authenticator|null
-     */
-    public function getAuthenticator(): ?Authenticator;
-
-    /**
-     * Authenticate the request with an authenticator.
-     *
-     * @param Authenticator $authenticator
-     * @return \Saloon\Http\PendingRequest
-     */
-    public function authenticate(Authenticator $authenticator): static;
-
-    /**
-     * Authenticate the request with an Authorization header.
-     *
-     * @param string $token
-     * @param string $prefix
-     * @return \Saloon\Http\PendingRequest
-     */
-    public function withTokenAuth(string $token, string $prefix = 'Bearer'): static;
-
-    /**
-     * Authenticate the request with "basic" authentication.
-     *
-     * @param string $username
-     * @param string $password
-     * @return \Saloon\Http\PendingRequest
-     */
-    public function withBasicAuth(string $username, string $password): static;
-
-    /**
-     * Authenticate the request with "digest" authentication.
-     *
-     * @param string $username
-     * @param string $password
-     * @param string $digest
-     * @return \Saloon\Http\PendingRequest
-     */
-    public function withDigestAuth(string $username, string $password, string $digest): static;
-
-    /**
-     * Authenticate the request with a query parameter token.
-     *
-     * @param string $parameter
-     * @param string $value
-     * @return \Saloon\Http\PendingRequest
-     */
-    public function withQueryAuth(string $parameter, string $value): static;
-
-    /**
-     * Invoke a callable where a given value returns a truthy value.
-     *
-     * @param mixed $value
-     * @param callable $callback
-     * @param callable|null $default
-     * @return \Saloon\Http\PendingRequest
-     */
-    public function when(mixed $value, callable $callback, callable $default = null): static;
-
-    /**
-     * Invoke a callable when a given value returns a falsy value.
-     *
-     * @param mixed $value
-     * @param callable $callback
-     * @param mixed $default
-     * @return \Saloon\Http\PendingRequest
-     */
-    public function unless(mixed $value, callable $callback, callable $default = null): static;
-
-    /**
-     * Access the config
-     *
-     * @return ArrayStore
-     */
-    public function config(): ArrayStore;
-
-    /**
-     * Access the headers
-     *
-     * @return ArrayStore
-     */
-    public function headers(): ArrayStore;
-
-    /**
-     * Access the middleware pipeline
-     *
-     * @return MiddlewarePipeline
-     */
-    public function middleware(): MiddlewarePipeline;
-
-    /**
-     * Specify a mock client.
-     *
-     * @param MockClient $mockClient
-     * @return \Saloon\Http\PendingRequest
-     */
-    public function withMockClient(MockClient $mockClient): static;
-
-    /**
-     * Get the mock client.
-     *
-     * @return MockClient|null
-     */
-    public function getMockClient(): ?MockClient;
-
-    /**
-     * Determine if the instance has a mock client
-     *
-     * @return bool
-     */
-    public function hasMockClient(): bool;
-
-    /**
-     * Access the query parameters
-     *
-     * @return ArrayStore
-     */
-    public function query(): ArrayStore;
-
     /**
      * Execute the response pipeline.
      *
